+++ Backend/src/ConsertaPraMim.Application/Interfaces/IAdminNoShowAuditService.cs
@@ -1,0 +1,8 @@
+using ConsertaPraMim.Application.DTOs;
+
+namespace ConsertaPraMim.Application.Interfaces;
+
+public interface IAdminNoShowAuditService
+{
+    Task<AdminNoShowAuditDto> GetAuditAsync(AdminNoShowAuditQueryDto query);
+}